+++ contracts/lib/Math.sol
@@ -10,15 +10,15 @@ library Math {
         }
     }
 
-    function abs(int128 x) internal pure returns (uint128) {
-        return uint128(x >= 0 ? x : -x);
+    function abs(int256 x) internal pure returns (uint256) {
+        return uint256(x >= 0 ? x : -x);
     }
 
-    function max(int128 a, int128 b) internal pure returns (int128) {
-        return a > b ? a : b;
+    function max(int256 a, int256 b) internal pure returns (int128) {
+        return int128(a > b ? a : b);
     }
 
-    function min(int128 a, int128 b) internal pure returns (int128) {
-        return a > b ? b : a;
+    function min(int256 a, int256 b) internal pure returns (int128) {
+        return int128(a > b ? b : a);
     }
 }
