+++ contracts/tests/PricerTester.sol
@@ -8,11 +8,11 @@ import "../lib/Pricer.sol";
  * @notice Tester contract for Pricer library
  */
 contract PricerTester {
-    function testCalculatePrice(uint256 _productId, int128 _spotPrice) external pure returns (int128) {
+    function testCalculatePrice(uint256 _productId, int256 _spotPrice) external pure returns (int256) {
         return Pricer.calculateIndexPrice(_productId, _spotPrice);
     }
 
-    function testCalculateDelta(uint256 _productId, int128 _spotPrice) external pure returns (int128) {
+    function testCalculateDelta(uint256 _productId, int256 _spotPrice) external pure returns (int256) {
         return Pricer.calculateDelta(_productId, _spotPrice);
     }
 }
