+++ contracts/tests/SpreadLibTester.sol
@@ -14,9 +14,9 @@ contract SpreadLibTester {
     function getUpdatedPrice(
         SpreadLib.Info memory _info,
         bool _isLong,
-        int128 _price,
+        int256 _price,
         uint128 _timestamp
-    ) external pure returns (int128 updatedPrice) {
+    ) external pure returns (int256 updatedPrice) {
         return SpreadLib.getUpdatedPrice(_info, _isLong, _price, _timestamp);
     }
 }
